app/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg==3.2.11
openai==1.44.1
sentence-transformers==3.1.1
numpy==1.26.4
